# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: amoCRM Admin API
  version: v1
  summary: Internal admin panel API.
  description: |
    Employee session API for the admin panel. Authenticate with the
    `admin_session` cookie. Mutating requests require `X-Requested-With: admin-ui`
    and `Origin` equal to `ADMIN_PUBLIC_ORIGIN`.
servers:
  - url: /
security:
  - adminSession: []
...
paths:
...
  /api/v1/system/employees:
    get:
      tags: [Employees]
      operationId: listEmployees
      summary: List employees.
      parameters:
        - $ref: '#/components/parameters/RequestID'
      responses:
        '200':
          description: Employee list.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeList'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Employees]
      operationId: createEmployee
      summary: Create an employee.
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/RequestedWith'
        - $ref: '#/components/parameters/Origin'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/system/employees/{id}:
    patch:
      tags: [Employees]
      operationId: patchEmployee
      summary: Update an employee. Role or status changes revoke sessions.
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/RequestedWith'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/ID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchEmployeeRequest'
      responses:
        '200':
          description: Employee updated.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/system/employees/{id}/sessions/revoke:
    post:
      tags: [Employees]
      operationId: revokeEmployeeSessions
      summary: Revoke all sessions of an employee.
      parameters:
        - $ref: '#/components/parameters/RequestID'
        - $ref: '#/components/parameters/RequestedWith'
        - $ref: '#/components/parameters/Origin'
        - $ref: '#/components/parameters/ID'
      responses:
        '200':
          description: Sessions revoked.
          headers:
            X-Request-ID:
              $ref: '#/components/headers/RequestID'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OK'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    adminSession:
      type: apiKey
      in: cookie
      name: admin_session
  parameters:
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
    RequestedWith:
      name: X-Requested-With
      in: header
      required: true
      description: Must equal `admin-ui` for POST, PATCH, PUT and DELETE.
      schema:
        type: string
        enum: [admin-ui]
    Origin:
      name: Origin
      in: header
      required: true
      description: Must equal `ADMIN_PUBLIC_ORIGIN` (trailing slash ignored).
      schema:
        type: string
        format: uri
    ID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  headers:
    RequestID:
      schema:
        type: string
        format: uuid
  schemas:
...
    CreateEmployeeRequest:
      type: object
      additionalProperties: false
      required: [email, name, role, password]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        password:
          type: string
          minLength: 8
    PatchEmployeeRequest:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/EmployeeStatus'
    Role:
      type: string
      enum: [viewer, operator, admin]
    EmployeeStatus:
      type: string
      enum: [active, disabled]
...
    Employee:
      type: object
      additionalProperties: false
      required: [id, email, name, role, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        name:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        status:
          $ref: '#/components/schemas/EmployeeStatus'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    EmployeeList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        next_cursor:
          type: [string, 'null']
        total:
          type: [integer, 'null']
...
    OK:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
              enum:
                - unauthenticated
                - forbidden
                - not_found
                - invalid_argument
                - conflict
                - rate_limited
                - internal
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
  responses:
    Unauthenticated:
      description: Authentication required or failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: CSRF check failed or permission denied.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    InvalidArgument:
      description: Request is invalid.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource already exists.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    InternalError:
      description: Internal error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'